move SimulationState and SimulationDataTracking to a separate file, simplify MeshNode constructor
To simplify the MeshNode constructor, we need (want) to import the types for
them. But they are defined in a file that imports lib/node.py so that would be
a circular import. Move them to a separate file so both
lib/discrete_event_sim.py and lib/node.py can use them. Then simplify the
MeshNode constructor to accept them and unpack them internally.
